refactor verification to use seperate processor
