+++ cutter-lib/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,2 @@
 org.x2a.cutter.processor.CutterProcessor
+org.x2a.cutter.processor.VerificationProcessor
